コード共通化
